Changed float types in obj_m x and y attributes as they contain the north/east values (doubles) and that caused a precision difference in the floating point values.

+++ include/obj.h
@@ -10,15 +10,15 @@
 namespace tracking{
 /*Object in meters*/
 struct obj_m{
-    float x     = 0;
-    float y     = 0;
+    double x     = 0;
+    double y     = 0;
     int frame   = -1;
     int cl      = -1;
     int w       = 0;
     int h       = 0;
 
     obj_m(){}
-    obj_m(const float x_, const float y_, const int frame_, const int cl_, const int width, const int height) : x(x_), y(y_), frame(frame_), cl(cl_), w(width), h(height) {}
+    obj_m(const double x_, const double y_, const int frame_, const int cl_, const int width, const int height) : x(x_), y(y_), frame(frame_), cl(cl_), w(width), h(height) {}
     void print();
 };
 }
